ErrorMessage refactor
